fix(preflight): keep the store's transport tests out of the unit matrix
The unit legs run with `--disable-socket --allow-unix-socket`, so the class
that proves the write really leaves the process — a real loopback server and
real httpx, no stub below the client — failed every leg with SocketBlockedError.

`@pytest.mark.allow_hosts(["127.0.0.1"])` is the repo's sanctioned opt-in and
fixed four of the five, but the fifth still failed reproducibly under `-n auto`
in a full-directory run while passing in isolation. That is the wrong argument
to keep having: a test that stands a server up and holds a connection open does
not belong in a suite that runs twelve matrix legs under xdist behind a socket
guard with a Windows carve-out.

So the class moves to tests/integration/, which runs single-threaded with a
120s timeout and reruns on transient network errors, and the unit file keeps the
property that actually matters — that `create_task` schedules without running,
so the caller returns before the write starts — as a deterministic assertion on
call ordering that never opens a socket.

Also two artefacts the new `Input.workflow_slug` field obliges. B006 requires a
new field on the public contract surface to be recorded in the contract ledger,
which adds it for the three extraction inputs that inherit it; and the capability
manifest is regenerated so the SDK-surface diff is visible in Files changed.

Verified under CI's own flags: unit 830 passed on three consecutive runs,
integration 5 passed.
