+++ tests/unit/execution/test_preflight_persist.py
@@ -3,7 +3,7 @@
 from __future__ import annotations
 
 import asyncio
-from contextlib import asynccontextmanager
+from unittest import mock
 
 import pytest
 
@@ -348,151 +348,59 @@ class TestTheWriteNeverFailsTheGate:
         )
 
 
-class TestTheWriteReallyLeavesTheProcess:
-    """No stubs below the HTTP client: a real socket, over real httpx.
+class TestTheWriteIsOffTheCallersPath:
+    """The ordering the whole design rests on, without a socket.
 
-    Every other test here substitutes ``post_check_result``, which is the right
-    seam for asserting *what* is sent — but it cannot show the send is genuinely
-    off the caller's path, because a stub returns immediately whether the real
-    thing would or not.
-
-    Each scenario runs under a hard deadline, so a regression that makes the
-    write blocking fails the test instead of hanging the suite.
+    The real-transport version of this lives in
+    ``tests/integration/test_preflight_persist.py``. This one is deterministic:
+    it never opens a connection, so it holds under the unit suite's socket
+    guard and under xdist.
     """
 
-    DEADLINE = 10.0
+    def test_the_caller_returns_before_the_write_even_starts(self):
+        # loop.create_task schedules without running, and persist_check_result
+        # awaits nothing after it — so the write's first tick cannot happen
+        # until the caller has already returned.
+        trace: list[str] = []
 
-    @staticmethod
-    def _run(scenario) -> None:
-        async def _bounded():
-            await asyncio.wait_for(
-                scenario(), timeout=TestTheWriteReallyLeavesTheProcess.DEADLINE
-            )
-
-        asyncio.run(_bounded())
-
-    @staticmethod
-    @asynccontextmanager
-    async def _serving(handler):
-        """A server on a free loopback port, torn down without waiting on handlers.
-
-        Handlers are tracked and cancelled rather than awaited: one of them holds
-        a connection open on purpose, and ``wait_closed`` would block on it.
-        """
-        live: set[asyncio.Task[None]] = set()
-
-        async def _tracked(reader, writer):
-            task = asyncio.current_task()
-            assert task is not None
-            live.add(task)
-            try:
-                await handler(reader, writer)
-            finally:
-                live.discard(task)
-                writer.close()
-
-        server = await asyncio.start_server(_tracked, "127.0.0.1", 0)
-        port = server.sockets[0].getsockname()[1]
-        try:
-            yield f"http://127.0.0.1:{port}/continuous-preflight/check-results"
-        finally:
-            for task in list(live):
-                task.cancel()
-            server.close()
-
-    @staticmethod
-    def _respond(status: bytes, body: bytes = b""):
-        async def _handler(reader, writer):
-            await reader.read(1)
-            writer.write(
-                b"HTTP/1.1 " + status + b"\r\nConnection: close\r\n"
-                b"Content-Type: application/json\r\n"
-                b"Content-Length: " + str(len(body)).encode() + b"\r\n\r\n" + body
-            )
-            await writer.drain()
-
-        return _handler
-
-    def _persist(self, endpoint: str, timeout: float = 5.0):
-        return persist.persist_check_result(
-            _gate(workflow_slug=SLUG), _verdict(), endpoint=endpoint, timeout=timeout
-        )
-
-    def test_the_caller_returns_before_the_store_answers(self):
-        # The point of the whole design: the app endpoint opens a Polaris catalog,
-        # may create the table and commits Parquet. Were this ever awaited, every
-        # gated run in the fleet would carry that latency.
-        connected = asyncio.Event()
-
-        async def _never_answer(reader, writer):
-            connected.set()
-            await asyncio.Event().wait()  # held open until cancelled
+        async def _slow(row, *, endpoint, timeout):
+            trace.append("write:started")
+            await asyncio.sleep(0.2)
+            trace.append("write:finished")
 
         async def scenario():
-            async with self._serving(_never_answer) as endpoint:
+            with mock.patch.object(persist, "post_check_result", _slow):
+                task = persist.persist_check_result(
+                    _gate(workflow_slug=SLUG),
+                    _verdict(),
+                    endpoint="http://store/x/check-results",
+                    timeout=30,
+                )
+                trace.append("caller:returned")
+                assert task is not None
+                await task
+            assert trace == ["caller:returned", "write:started", "write:finished"]
+
+        asyncio.run(scenario())
+
+    def test_a_slow_write_costs_the_caller_nothing(self):
+        async def _slow(row, *, endpoint, timeout):
+            await asyncio.sleep(0.2)
+
+        async def scenario():
+            with mock.patch.object(persist, "post_check_result", _slow):
                 loop = asyncio.get_running_loop()
                 started = loop.time()
-                task = self._persist(endpoint, timeout=30)
-                assert loop.time() - started < 0.05
-                assert task is not None and not task.done()
-
-                # The request does reach the socket, and is still in flight long
-                # after the caller moved on.
-                await connected.wait()
-                assert not task.done()
-
-                task.cancel()
-                await asyncio.gather(task, return_exceptions=True)
-
-        self._run(scenario)
-
-    def test_a_store_that_refuses_the_connection_does_not_reach_the_caller(self):
-        async def scenario():
-            async with self._serving(self._respond(b"200 OK")) as endpoint:
-                pass  # the port is free again, and nothing is listening on it
-            task = self._persist(endpoint, timeout=2)
-            assert task is not None
-            # Awaiting re-raises what the task caught, which no production caller
-            # does — the guarantee under test is that scheduling it did not raise.
-            await asyncio.gather(task, return_exceptions=True)
-            assert isinstance(task.exception(), Exception)
-
-        self._run(scenario)
-
-    def test_a_rejected_row_does_not_reach_the_caller(self):
-        async def scenario():
-            async with self._serving(self._respond(b"422 Unprocessable Entity")) as e:
-                task = self._persist(e)
+                task = persist.persist_check_result(
+                    _gate(workflow_slug=SLUG),
+                    _verdict(),
+                    endpoint="http://store/x/check-results",
+                    timeout=30,
+                )
+                elapsed = loop.time() - started
                 assert task is not None
                 await task
-                assert task.exception() is None
+            # Two orders of magnitude under the write it scheduled.
+            assert elapsed < 0.02, f"scheduling cost {elapsed:.4f}s"
 
-        self._run(scenario)
-
-    def test_a_broken_store_does_not_reach_the_caller(self):
-        async def scenario():
-            async with self._serving(self._respond(b"503 Service Unavailable")) as e:
-                task = self._persist(e)
-                assert task is not None
-                await task
-                assert task.exception() is None
-
-        self._run(scenario)
-
-    def test_a_rejection_never_echoes_the_verdict_into_a_log(self, caplog):
-        # A 4xx body is FastAPI's validation error, which renders the input it
-        # rejected — and the input is the verdict.
-        body = b'{"detail":"invalid payload: db://svc:hunter2@example-host:3306"}'
-
-        async def scenario():
-            async with self._serving(
-                self._respond(b"422 Unprocessable Entity", body)
-            ) as endpoint:
-                task = self._persist(endpoint)
-                assert task is not None
-                await task
-
-        with caplog.at_level("DEBUG"):
-            self._run(scenario)
-        assert "hunter2" not in caplog.text
-        assert "422" in caplog.text  # the status is what does get reported
+        asyncio.run(scenario())
